requirements.txt
# RAG Chunking & Embedding Strategy Evaluation Project
# Python Dependencies

# Core Data Processing
numpy>=1.21.0
pandas>=1.3.0

# Vector Database
qdrant-client>=1.16.0,<2.0.0

# Embedding Models & NLP
sentence-transformers>=2.2.0
torch>=1.9.0

# Document Parsing
PyPDF2>=3.0.0
python-docx>=0.8.11
openpyxl>=3.0.0
python-pptx>=0.6.21

# Utilities
requests>=2.28.0
tqdm>=4.62.0
